fix: move isCategoryLocked out of use server file; fix Supabase join type casts
